sitemap generation: - use app/sitemap.xml/routes.ts for sitemap generation, to avoid throwing error in dev - generates top level URLs dynamically
